res made the following changes: - Added CS_FUNCTION_NAME macro, gives name of current function. - Added support for GL_GREMEDY_string_marker extension. The GL   renderer uses it to write some out some info. Helpful to link   GL operations to CS methods in a GL trace. - Made some cleanups & simplifications to the GL render-to-texture   code. - Made some cleanups & simplifications to the GL renderer and   cavases in general - mostly cut down redundant matrix and   viewport setups. - Fixed returning of a stack var in minidump.cpp. - Changed csFontCache to store the size of known fonts in floats. - Fixed a csString passed through '...' in csfgen (really). - Fixed gfxtest at some place were a 0 char* could be passed to a   function that expected a non-0 char*.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20853 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csplugincommon/canvas/fontcache.h
@@ -144,7 +144,7 @@ public:
   {
     iFont* font;
     /// The font size this font was cached for
-    int fontSize;
+    float fontSize;
     PlaneGlyphsArray planeGlyphs;
   };
   /// the current clipping rect
